refactor: Extract variable assertions
Move the variable assertion logic into a new class to separate the different concerns.
